Stable pulled from original repos

+++ app/controllers/communication_channels_controller.rb
@@ -23,60 +23,74 @@
 #
 # In this API, the `:user_id` parameter can always be replaced with `self` if
 # the requesting user is asking for his/her own information.
+#
+# @object Communication Channel
+#     {
+#       // The ID of the communication channel.
+#       "id": 16,
+#
+#       // The address, or path, of the communication channel.
+#       "address": "sheldon@caltech.example.com",
+#
+#       // The type of communcation channel being described. Possible values
+#       // are: "email", "sms", "chat", "facebook" or "twitter". This field
+#       // determines the type of value seen in "address".
+#       "type": "email",
+#
+#       // The position of this communication channel relative to the user's
+#       // other channels when they are ordered.
+#       "position": 1,
+#
+#       // The ID of the user that owns this communication channel.
+#       "user_id": 1,
+#
+#       // The current state of the communication channel. Possible values are:
+#       // "unconfirmed" or "active".
+#       "workflow_state": "active"
+#     }
 class CommunicationChannelsController < ApplicationController
   before_filter :require_user, :only => [:create, :destroy]
   before_filter :reject_student_view_student
 
   include Api::V1::CommunicationChannel
 
-  # @API
-  # List the communication channels for a user.
+  # @API List user communication channels
+  #
+  # Returns a list of communication channels for the specified user, sorted by
+  # position.
   #
   # @example_request
-  #   curl https://<canvas>/api/v1/users/12345/communication_channels -H 'Authorization: Bearer <ACCESS_TOKEN>'
+  #     curl https://<canvas>/api/v1/users/12345/communication_channels \ 
+  #          -H 'Authorization: Bearer <token>'
   #
-  # @example_response
-  #   [
-  #     { "id": 1, "address": "bieberfever@example.com", "type": "email", "position": 1, "user_id": 12345 },
-  #     { "id": 2, "address": "8018675309", "type": "sms", "position": 2, "user_id": 12345 }
-  #   ]
+  # @returns [Communication Channel]
   def index
     @user = api_find(User, params[:user_id])
     return unless authorized_action(@user, @current_user, :read)
 
     channels = Api.paginate(@user.communication_channels.unretired, self,
-      api_v1_communication_channels_path).map do |cc|
+      api_v1_communication_channels_url).map do |cc|
         communication_channel_json(cc, @current_user, session)
       end
 
     render :json => channels
   end
 
-  # @API Create a new communication channel.
+  # @API Create a communication channel
+  #
+  # Creates a new communication channel for the specified user.
   #
   # @argument communication_channel[address] An email address or SMS number.
   # @argument communication_channel[type] [email|sms] The type of communication channel.
-  # @argument skip_confirmation [Optional] Only valid for site admins making requests; If '1', the
-  # channel is automatically validated and no confirmation email or SMS is sent. Otherwise, the user must
-  # respond to a confirmation message to confirm the channel.
+  # @argument skip_confirmation [Optional] Only valid for site admins making requests; If '1', the channel is automatically validated and no confirmation email or SMS is sent. Otherwise, the user must respond to a confirmation message to confirm the channel.
   #
   # @example_request
+  #     curl https://<canvas>/api/v1/users/1/communication_channels \ 
+  #          -H 'Authorization: Bearer <token>' \ 
+  #          -d 'communication_channel[address]=new@example.com' \ 
+  #          -d 'communication_channel[type]=email' \ 
   #
-  #   curl https://<canvas>/api/v1/users/1/communication_channels \ 
-  #     -H 'Authorization: Bearer <ACCESS_TOKEN>' \ 
-  #     -d 'communication_channel[address]=new@example.com' \ 
-  #     -d 'communication_channel[type]=email' \ 
-  #
-  # @example_response
-  #
-  #   {
-  #     "id": 2,
-  #     "address": "new@example.com",
-  #     "type": "email",
-  #     "workflow_state": "unconfirmed",
-  #     "user_id": 1,
-  #     "position": 2
-  #   }
+  # @returns Communication Channel
   def create
     @user = api_request? ? api_find(User, params[:user_id]) : @current_user
 
@@ -127,8 +141,8 @@ class CommunicationChannelsController < ApplicationController
   end
 
   def confirm
-    nonce = params[:nonce]
-    cc = CommunicationChannel.unretired.find_by_confirmation_code(nonce)
+    @nonce = params[:nonce]
+    cc = CommunicationChannel.unretired.find_by_confirmation_code(@nonce)
     @headers = false
     if cc
       @communication_channel = cc
@@ -139,7 +153,7 @@ class CommunicationChannelsController < ApplicationController
       @root_account ||= @user.pseudonyms.first.try(:account) if @user.pre_registered?
       @root_account ||= @user.enrollments.first.try(:root_account) if @user.creation_pending?
       unless @root_account
-        account = @user.account_users.first.try(:account)
+        account = @user.accounts.first
         @root_account = account.try(:root_account)
       end
       @root_account ||= @domain_root_account
@@ -166,11 +180,8 @@ class CommunicationChannelsController < ApplicationController
       end
 
       # load merge opportunities
-      other_ccs = CommunicationChannel.active.by_path(cc.path).of_type(cc.path_type).find(:all, :conditions => ["communication_channels.id<>?", cc.id], :include => :user)
-      merge_users = (other_ccs.map(&:user)).uniq
+      merge_users = cc.merge_candidates
       merge_users << @current_user if @current_user && !@user.registered? && !merge_users.include?(@current_user)
-      User.send(:preload_associations, merge_users, { :pseudonyms => :account })
-      merge_users.reject! { |u| u != @current_user && u.pseudonyms.all? { |p| p.deleted? } }
       # remove users that don't have a pseudonym for this account, or one can't be created
       merge_users = merge_users.select { |u| u.find_or_initialize_pseudonym_for_account(@root_account, @domain_root_account) }
       @merge_opportunities = []
@@ -180,8 +191,7 @@ class CommunicationChannelsController < ApplicationController
         if root_account_pseudonym
           @merge_opportunities << [user, [root_account_pseudonym]]
         else
-          user.pseudonyms.each do |p|
-            next unless p.active?
+          user.all_active_pseudonyms.each do |p|
             # populate reverse association
             p.user = user
             (account_to_pseudonyms_hash[p.account] ||= []) << p
@@ -194,21 +204,32 @@ class CommunicationChannelsController < ApplicationController
       end
       @merge_opportunities.sort! { |a, b| [a.first == @current_user ? 0 : 1, a.first.name] <=> [b.first == @current_user ? 0 : 1, b.first.name] }
 
+      js_env :PASSWORD_POLICY => @domain_root_account.password_policy
+
       if @current_user && params[:confirm].present? && @merge_opportunities.find { |opp| opp.first == @current_user }
-        cc.confirm
-        @enrollment.accept if @enrollment
-        @user.move_to_user(@current_user) if @user != @current_user
-        # create a new pseudonym if necessary and possible
-        pseudonym = @current_user.find_or_initialize_pseudonym_for_account(@root_account, @domain_root_account)
-        pseudonym.save! if pseudonym && pseudonym.changed?
+        @user.transaction do
+          @current_user.transaction do
+            cc.confirm
+            @enrollment.accept if @enrollment
+            UserMerge.from(@user).into(@current_user) if @user != @current_user
+            # create a new pseudonym if necessary and possible
+            pseudonym = @current_user.find_or_initialize_pseudonym_for_account(@root_account, @domain_root_account)
+            pseudonym.save! if pseudonym && pseudonym.changed?
+          end
+        end
       elsif @current_user && @current_user != @user && @enrollment && @user.registered?
+
         if params[:transfer_enrollment].present?
-          cc.active? || cc.confirm
-          @enrollment.user = @current_user
-          # accept will save it
-          @enrollment.accept
-          @user.touch
-          @current_user.touch
+          @user.transaction do
+            @current_user.transaction do
+              cc.active? || cc.confirm
+              @enrollment.user = @current_user
+              # accept will save it
+              @enrollment.accept
+              @user.touch
+              @current_user.touch
+            end
+          end
         else
           # render
           return
@@ -224,9 +245,9 @@ class CommunicationChannelsController < ApplicationController
       else
         # Open registration and admin-created users are pre-registered, and have already claimed a CC, but haven't
         # set up a password yet
-        @pseudonym = @user.pseudonyms.active.find(:first, :conditions => {:password_auto_generated => true, :account_id => @root_account.id} ) if @user.pre_registered? || @user.creation_pending?
+        @pseudonym = @root_account.pseudonyms.active.where(:password_auto_generated => true, :user_id => @user).first if @user.pre_registered? || @user.creation_pending?
         # Users implicitly created via course enrollment or account admin creation are creation pending, and don't have a pseudonym yet
-        @pseudonym ||= @user.pseudonyms.build(:account => @root_account, :unique_id => cc.path) if @user.creation_pending?
+        @pseudonym ||= @root_account.pseudonyms.build(:user => @user, :unique_id => cc.path) if @user.creation_pending?
         # We create the pseudonym with unique_id = cc.path, but if that unique_id is taken, just nil it out and make the user come
         # up with something new
         @pseudonym.unique_id = '' if @pseudonym && @pseudonym.new_record? && @root_account.pseudonyms.active.custom_find_by_unique_id(@pseudonym.unique_id)
@@ -239,6 +260,7 @@ class CommunicationChannelsController < ApplicationController
           if Canvas.redis_enabled? && @merge_opportunities.length == 1
             Canvas.redis.rpush('single_user_registered_new_account_stats', {:user_id => @user.id, :registered_at => Time.now.utc }.to_json)
           end
+          @user.require_acceptance_of_terms = require_terms?
           @user.attributes = params[:user]
           @pseudonym.attributes = params[:pseudonym]
           @pseudonym.communication_channel = cc
@@ -247,7 +269,14 @@ class CommunicationChannelsController < ApplicationController
           @pseudonym.require_password = true
           @pseudonym.password_confirmation = @pseudonym.password = params[:pseudonym][:password] if params[:pseudonym]
 
-          return unless @pseudonym.valid?
+          unless @pseudonym.valid? && @user.valid?
+            return render :json => {
+                            :errors => {
+                              :user => @user.errors.as_json[:errors],
+                              :pseudonym => @pseudonym.errors.as_json[:errors]
+                            }
+                          }, :status => :bad_request
+          end
 
           # They may have switched e-mail address when they logged in; create a CC if so
           if @pseudonym.unique_id != cc.path
@@ -292,7 +321,7 @@ class CommunicationChannelsController < ApplicationController
       flash[:notice] = t 'notices.registration_confirmed', "Registration confirmed!"
       respond_to do |format|
         format.html { @enrollment ? redirect_to(course_url(@course)) : redirect_back_or_default(dashboard_url) }
-        format.json { render :json => cc.to_json(:except => [:confirmation_code] ) }
+        format.json { render :json => {:url => @enrollment ? course_url(@course) : dashboard_url} }
       end
     end
   end
@@ -310,28 +339,24 @@ class CommunicationChannelsController < ApplicationController
     render :json => {:re_sent => true}
   end
 
-  # @API Delete a communication channel.
+  # @API Delete a communication channel
+  #
   # Delete an existing communication channel.
   #
   # @example_request
-  #   curl https://<canvas>/api/v1/users/5/communication_channels/3
-  #     -H 'Authorization: Bearer <ACCESS_TOKEN>
-  #     -X DELETE
+  #     curl https://<canvas>/api/v1/users/5/communication_channels/3
+  #          -H 'Authorization: Bearer <token>
+  #          -X DELETE
   #
-  # @example_response
-  #   {
-  #     "id": 3,
-  #     "address": "new@example.com",
-  #     "type": "email",
-  #     "workflow_state": "deleted",
-  #     "user_id": 5,
-  #     "position": 2
-  #   }
+  # @returns Communication Channel
   def destroy
     @user = api_request? ? api_find(User, params[:user_id]) : @current_user
     @cc   = @user.communication_channels.find(params[:id]) if params[:id]
 
     return render_unauthorized_action unless has_api_permissions?
+    if @cc.imported? && !@domain_root_account.edit_institution_email?
+      return render_unauthorized_action
+    end
 
     if @cc.nil? || @cc.destroy
       @user.touch
@@ -350,4 +375,10 @@ class CommunicationChannelsController < ApplicationController
     @user == @current_user ||
       @user.grants_right?(@current_user, session, :manage_user_details)
   end
+
+  def require_terms?
+    # a plugin could potentially set this
+    @require_terms
+  end
+  helper_method :require_terms?
 end
